Added slug exceptions
